Update tests.py
Very interesting problem I just spent a few hours on this! I think the first extra test is wrong. 22 time at the beginning, after first hourglass expires (-13) there is only 9 left so the max number of turns should be 9 (which is the answer my algorithm calculated) but the answer in the test was 16.